requirements.txt
asgiref==3.8.1
Django==5.0.6
django-easy-validation==0.1
django-rest-framework==0.1.0
djangorestframework==3.15.1
pendulum==3.0.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
regex==2024.5.10
six==1.16.0
sqlparse==0.5.0
time-machine==2.14.1
tzdata==2024.1
